getPythonPath() has been rewritten and is now called getPathToPyConfig()
getPathToPyConfig() returns the path to the configuration files that come
with openWNS. They are located in a directory next to the executable which
is called lib/PyConfig. To find this directory, the application reads the
value of the link /proc/self/exe which points to the running executable and
modifies this path accordingly. This (with a patch on the build support)
allows to move "sandboxes" around, since no absolute paths are compiled into
the simulator any more.